refactor: rename public dashboard wire identifiers to workspace
Rename catalog-link contract file, capability ids, GraphQL pin operation
documents, pin-error query param, view-mode storage key, and contract
prose so product language matches /workspace without keeping aliases.
