+++ docs/contracts/_product.json
@@ -5,7 +5,12 @@
     "controlled REST / GraphQL / MCP surface",
     "admin maintenance surface"
   ],
-  "surfaces": ["web", "rest", "graphql", "mcp"],
+  "surfaces": [
+    "web",
+    "rest",
+    "graphql",
+    "mcp"
+  ],
   "roles": {
     "anon": "Unauthenticated visitor; can browse all public information.",
     "user": "Authenticated, non-suspended user; can manage personal state and perform social actions.",
