+++ docs/contracts/_ui.json
@@ -10,7 +10,7 @@
   "Model Property Priority - Tertiary": "Tertiary properties support audit or exact reconstruction: internal IDs, raw timestamps, platform metadata, long nested relations, and repeated parent objects. Hide or de-emphasize them in ordinary UI; MCP default output may omit, redact, or compact them.",
   "Model Property Priority - Surfaces": "REST may expose fuller records for application clients. MCP default output preserves primary and important secondary fields, simplifies repeated low-priority fields, and reserves exact raw payloads for explicit full mode where supported; the deprecated summary input is an alias for default.",
   "Basic Info Card": "* Used in detail page sidebars or collapsible panels.\n* Displays structured fields; does not carry comment or homework interactions.",
-  "Context Tabs": "* Not used for course, section, or teacher catalog detail pages (see Detail Two-Column Stream).\n* May still appear on workspace surfaces that switch distinct task contexts (for example dashboard calendar/homework branches).\n* When present, context navigation is a horizontal, touch-scrollable row with an overflow affordance on mobile and a vertical sidebar on larger screens; the active destination remains programmatically identified and scrolls into view.",
+  "Context Tabs": "* Not used for course, section, or teacher catalog detail pages (see Detail Two-Column Stream).\n* May still appear on workspace surfaces that switch distinct task contexts (for example workspace calendar/homework branches).\n* When present, context navigation is a horizontal, touch-scrollable row with an overflow affordance on mobile and a vertical sidebar on larger screens; the active destination remains programmatically identified and scrolls into view.",
   "Global Search Results": "* Catalog group order is sections → teachers → courses → links, then signed-in homeworks and todos.\n* Section items use title `course.namePrimary · teachers` (fall back to course name · section code when teachers are empty) and description `semester · campus · section.code` with code last.\n* Course items use title `course.namePrimary` and description `course.code`.\n* Teacher items use title `teacher.namePrimary` and description department or code.",
   "Calendar Export Dialog": "* Used to display single-object iCal and personal subscription iCal.\n* Presented as a copyable link; users are not required to manually construct the URL.",
   "Shell Layout": "* Desktop sidebar and main content scroll independently.\n* Signed-in desktop navigation keeps task links in Workspace; Explore and Admin tools are progressive groups.\n* A destination row is either a link or a disclosure, never both.\n* Signed-in mobile navigation fixes Today, Calendar, Tasks, and Explore above content and reserves its height and safe area.\n* Language and theme stay in the topbar; signed-in profile actions stay in the sidebar footer.\n* Anonymous topbars retain Sign in.\n* Mobile topbar/navigation controls are at least 44 by 44 CSS pixels.\n* Selected system/light/dark themes apply before first paint; system follows live OS changes with a no-JavaScript fallback.",
@@ -18,8 +18,8 @@
   "Workspace Empty States": "* Filter-caused empty results explain the active filter and provide a direct clear-filter action.\n* Setup-caused empty states pair their explanation with the existing setup or create action.\n* Routine zero-item states remain compact and do not grow merely to fill the workspace.",
   "Settings Navigation": "* Settings use concise labels in a desktop rail and scrollable mobile row.\n* `/account/settings/profile` is the default; legacy tab inputs only redirect.\n* Legacy appearance, language, and preferences values redirect to `/account/settings/preferences`.\n* Preferences reuse the pre-paint/localStorage theme mechanism and `/api/account/preferences`; URLs have no locale segments.\n* Authorizations remain distinct from linked sign-in accounts and require confirmation before revocation.\n* The active section owns its copy, exposes one aria-current page link, and scrolls into view on mobile.\n* Danger stays separate from routine settings.",
   "Contribution Heatmap": "* Public profile contribution history keeps a readable minimum cell size and scrolls horizontally when needed.\n* Each contribution day exposes its localized date/count to keyboard, touch, and assistive-technology users.",
-  "Dashboard Page Identity": "* Every signed-in dashboard branch exposes exactly one localized level-one heading.\n* The main content landmark uses the same localized branch label.\n* The document title identifies the active branch before the Life@USTC suffix.\n* Identity follows the semantic dashboard path after root or query compatibility redirects.",
-  "Dashboard Mobile Priority": "* Personal dashboard mobile layouts prioritize the next actionable information before directory-style shortcuts or wide planning grids.\n* Calendar uses a readable agenda at mobile widths while retaining semester, month, and week grids on desktop.\n* Homework, todo, and exam mobile toolbars keep status filters directly available, move card/list display mode into an accessible overflow menu, and keep primary controls at least 44 by 44 CSS pixels.",
+  "Workspace Page Identity": "* Every signed-in workspace branch exposes exactly one localized level-one heading.\n* The main content landmark uses the same localized branch label.\n* The document title identifies the active branch before the Life@USTC suffix.\n* Identity follows the semantic workspace path after root or query compatibility redirects.",
+  "Workspace Mobile Priority": "* Personal workspace mobile layouts prioritize the next actionable information before directory-style shortcuts or wide planning grids.\n* Calendar uses a readable agenda at mobile widths while retaining semester, month, and week grids on desktop.\n* Homework, todo, and exam mobile toolbars keep status filters directly available, move card/list display mode into an accessible overflow menu, and keep primary controls at least 44 by 44 CSS pixels.",
   "Social Sharing Metadata": "* Every rendered page exposes one localized description plus one complete Open Graph and Twitter card set.\n* Canonical and Open Graph URLs are absolute, same-origin, and exclude query strings and fragments.\n* Course, section, and teacher detail subroutes canonicalize to the entity root and use controlled localized summaries rather than editable user content.\n* Social images use the 1200 by 630 Life@USTC PNG card with explicit dimensions, MIME type, and localized alternative text.",
   "Localized Count Copy": "* User-visible English count copy distinguishes singular from plural forms; Chinese may share one count template when its grammar does not change.",
   "Navigation Landmarks": "* The desktop application sidebar is a real navigation landmark labelled Primary navigation / 主导航.\n* Signed-in mobile layouts expose distinct Mobile primary navigation / 移动主导航 and Secondary navigation / 次级导航 landmarks.\n* Each navigation surface exposes at most one aria-current page destination.\n* A localized skip link is the first focusable shell control and moves keyboard focus to the main landmark.\n* Footer links use one distinct Footer navigation / 页脚导航 landmark so assistive technology can distinguish the regions."
