+++ docs/contracts/admin.json
@@ -29,7 +29,9 @@
       "title": "Admin Entry",
       "auth": "admin",
       "web": {
-        "pages": ["/admin"]
+        "pages": [
+          "/admin"
+        ]
       },
       "display": {
         "fields": [
@@ -42,7 +44,9 @@
       "title": "Moderation",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/moderation"]
+        "pages": [
+          "/admin/moderation"
+        ]
       },
       "rest": {
         "routes": [
@@ -107,7 +111,9 @@
       "title": "User Management",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/users"]
+        "pages": [
+          "/admin/users"
+        ]
       },
       "rest": {
         "routes": [
@@ -119,7 +125,9 @@
             "path": "/api/admin/users/[id]",
             "method": "PATCH",
             "returns": "{ user: User }",
-            "notes": ["Returns 404 when the target user is not found."]
+            "notes": [
+              "Returns 404 when the target user is not found."
+            ]
           }
         ]
       },
@@ -139,7 +147,9 @@
       "title": "OAuth Client Management",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/oauth"]
+        "pages": [
+          "/admin/oauth"
+        ]
       },
       "rest": "unavailable",
       "mcp": "unavailable",
@@ -161,7 +171,9 @@
       "title": "Bus Management",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/bus"]
+        "pages": [
+          "/admin/bus"
+        ]
       },
       "rest": "unavailable",
       "mcp": "unavailable",
@@ -181,7 +193,11 @@
     "audit-log": {
       "title": "Audit Log",
       "auth": "admin",
-      "web": { "pages": ["/admin/audit"] },
+      "web": {
+        "pages": [
+          "/admin/audit"
+        ]
+      },
       "rest": "unavailable",
       "mcp": "unavailable",
       "display": {
@@ -208,7 +224,11 @@
     "aggregate-analytics": {
       "title": "Aggregate Analytics",
       "auth": "admin",
-      "web": { "pages": ["/admin/analytics"] },
+      "web": {
+        "pages": [
+          "/admin/analytics"
+        ]
+      },
       "rest": "unavailable",
       "mcp": "unavailable",
       "display": {
