+++ docs/contracts/comment.json
@@ -26,10 +26,14 @@
         "mutations": [
           {
             "name": "commentCreate",
-            "arguments": { "input": "CreateCommentInput!" },
+            "arguments": {
+              "input": "CreateCommentInput!"
+            },
             "returns": "CommentMutationPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "POST /api/community/comments",
             "mcp_equivalent": "community_comment_create"
           },
@@ -41,25 +45,35 @@
             },
             "returns": "CommentMutationPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "PATCH /api/community/comments/[id]",
             "mcp_equivalent": "community_comment_update"
           },
           {
             "name": "commentDelete",
-            "arguments": { "id": "ID!" },
+            "arguments": {
+              "id": "ID!"
+            },
             "returns": "DeleteMutationPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "DELETE /api/community/comments/[id]",
             "mcp_equivalent": "community_comment_delete"
           },
           {
             "name": "commentsDelete",
-            "arguments": { "ids": "[ID!]!" },
+            "arguments": {
+              "ids": "[ID!]!"
+            },
             "returns": "CommentDeleteBatchPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "DELETE /api/community/comments/batch",
             "mcp_equivalent": "graphql_operation_run",
             "notes": [
@@ -76,7 +90,9 @@
             },
             "returns": "CommentReactionMutationPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "POST /api/community/comments/[id]/reactions",
             "mcp_equivalent": "community_comment_reaction_add"
           },
@@ -88,7 +104,9 @@
             },
             "returns": "CommentReactionMutationPayload!",
             "auth": "user",
-            "required_scopes": ["comment:write"],
+            "required_scopes": [
+              "comment:write"
+            ],
             "rest_equivalent": "DELETE /api/community/comments/[id]/reactions",
             "mcp_equivalent": "community_comment_reaction_remove"
           }
@@ -268,7 +286,9 @@
       "title": "Comment Governance",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/moderation"]
+        "pages": [
+          "/admin/moderation"
+        ]
       },
       "rest": {
         "routes": [
@@ -316,10 +336,15 @@
       "title": "Comment Auxiliary Pages",
       "auth": "anon",
       "web": {
-        "pages": ["/community/comments/[id]", "/guides/markdown-support"]
+        "pages": [
+          "/community/comments/[id]",
+          "/guides/markdown-support"
+        ]
       },
       "display": {
-        "fields": ["Comment guidelines/help pages"]
+        "fields": [
+          "Comment guidelines/help pages"
+        ]
       }
     }
   }
