+++ docs/contracts/course.json
@@ -25,11 +25,17 @@
       "title": "Course List & Filter",
       "auth": "anon",
       "links": {
-        "models": ["Course"],
-        "ui": ["List Table"]
+        "models": [
+          "Course"
+        ],
+        "ui": [
+          "List Table"
+        ]
       },
       "web": {
-        "pages": ["/catalog/courses"]
+        "pages": [
+          "/catalog/courses"
+        ]
       },
       "rest": {
         "routes": [
@@ -81,18 +87,31 @@
           "course.category.namePrimary",
           "course.classType.namePrimary"
         ],
-        "filters": ["search", "educationLevelId", "categoryId", "classTypeId"]
+        "filters": [
+          "search",
+          "educationLevelId",
+          "categoryId",
+          "classTypeId"
+        ]
       }
     },
     "course-detail": {
       "title": "Course Detail",
       "auth": "anon",
       "links": {
-        "models": ["Course"],
-        "ui": ["Detail Hero", "Detail Two-Column Stream", "Basic Info Card"]
+        "models": [
+          "Course"
+        ],
+        "ui": [
+          "Detail Hero",
+          "Detail Two-Column Stream",
+          "Basic Info Card"
+        ]
       },
       "web": {
-        "pages": ["/catalog/courses/[jwId]"]
+        "pages": [
+          "/catalog/courses/[jwId]"
+        ]
       },
       "rest": {
         "routes": [
@@ -110,7 +129,9 @@
           {
             "name": "course",
             "parent": "Catalog",
-            "arguments": { "jwId": "Int!" },
+            "arguments": {
+              "jwId": "Int!"
+            },
             "returns": "Course",
             "rest_equivalent": "GET /api/catalog/courses/[jwId]",
             "mcp_equivalent": "catalog_course_get"
