+++ docs/contracts/description.json
@@ -21,10 +21,14 @@
         "mutations": [
           {
             "name": "descriptionSet",
-            "arguments": { "input": "UpsertDescriptionInput!" },
+            "arguments": {
+              "input": "UpsertDescriptionInput!"
+            },
             "returns": "DescriptionMutationPayload!",
             "auth": "user",
-            "required_scopes": ["description:write"],
+            "required_scopes": [
+              "description:write"
+            ],
             "rest_equivalent": "POST /api/community/descriptions",
             "mcp_equivalent": "community_description_set",
             "notes": [
@@ -92,7 +96,9 @@
       "title": "Description Governance",
       "auth": "admin",
       "web": {
-        "pages": ["/admin/moderation"]
+        "pages": [
+          "/admin/moderation"
+        ]
       },
       "rest": {
         "routes": [
