+++ docs/contracts/ical.json
@@ -19,10 +19,14 @@
       "title": "Section Calendar Dialog",
       "auth": "anon",
       "links": {
-        "ui": ["Calendar Export Dialog"]
+        "ui": [
+          "Calendar Export Dialog"
+        ]
       },
       "web": {
-        "pages": ["/catalog/sections/[jwId]"]
+        "pages": [
+          "/catalog/sections/[jwId]"
+        ]
       },
       "rest": {
         "routes": [
@@ -48,11 +52,18 @@
       "title": "Personal Calendar Subscription",
       "auth": "user",
       "links": {
-        "models": ["User", "Section"],
-        "ui": ["Calendar Export Dialog"]
+        "models": [
+          "User",
+          "Section"
+        ],
+        "ui": [
+          "Calendar Export Dialog"
+        ]
       },
       "web": {
-        "pages": ["/workspace/calendar"]
+        "pages": [
+          "/workspace/calendar"
+        ]
       },
       "rest": {
         "routes": [
@@ -87,7 +98,9 @@
         ]
       },
       "display": {
-        "fields": ["iCalendar URL copy button"]
+        "fields": [
+          "iCalendar URL copy button"
+        ]
       }
     },
     "multi-section-ical": {
