+++ docs/contracts/semester.json
@@ -16,7 +16,9 @@
       "title": "Semester List",
       "auth": "anon",
       "links": {
-        "models": ["Semester"]
+        "models": [
+          "Semester"
+        ]
       },
       "web": "stable",
       "rest": {
@@ -35,7 +37,9 @@
           {
             "name": "semesters",
             "parent": "Catalog",
-            "arguments": { "page": "PageInput" },
+            "arguments": {
+              "page": "PageInput"
+            },
             "returns": "SemesterPage!",
             "rest_equivalent": "GET /api/catalog/semesters",
             "mcp_equivalent": "catalog_semester_list",
@@ -72,7 +76,9 @@
       "title": "Current Semester",
       "auth": "anon",
       "links": {
-        "models": ["Semester"]
+        "models": [
+          "Semester"
+        ]
       },
       "rules": {
         "latest-startdate-wins": "The current semester is determined by date range; when multiple semester ranges overlap the current date, the semester with the latest startDate is selected to prevent an older range or import residue from shadowing a more specific newer semester."
@@ -83,7 +89,9 @@
           {
             "path": "/api/catalog/semesters/current",
             "returns": "Semester | 404",
-            "notes": ["Returns the current semester or 404 when none is found."]
+            "notes": [
+              "Returns the current semester or 404 when none is found."
+            ]
           }
         ]
       },
