+++ docs/contracts/teacher.json
@@ -20,11 +20,17 @@
       "title": "Teacher List & Filter",
       "auth": "anon",
       "links": {
-        "models": ["Teacher"],
-        "ui": ["List Table"]
+        "models": [
+          "Teacher"
+        ],
+        "ui": [
+          "List Table"
+        ]
       },
       "web": {
-        "pages": ["/catalog/teachers"]
+        "pages": [
+          "/catalog/teachers"
+        ]
       },
       "rest": {
         "routes": [
@@ -76,18 +82,29 @@
           "teacher.email",
           "teacher._count.sections"
         ],
-        "filters": ["search", "departmentId"]
+        "filters": [
+          "search",
+          "departmentId"
+        ]
       }
     },
     "teacher-detail": {
       "title": "Teacher Detail",
       "auth": "anon",
       "links": {
-        "models": ["Teacher"],
-        "ui": ["Detail Hero", "Detail Two-Column Stream", "Basic Info Card"]
+        "models": [
+          "Teacher"
+        ],
+        "ui": [
+          "Detail Hero",
+          "Detail Two-Column Stream",
+          "Basic Info Card"
+        ]
       },
       "web": {
-        "pages": ["/catalog/teachers/[id]"]
+        "pages": [
+          "/catalog/teachers/[id]"
+        ]
       },
       "rest": {
         "routes": [
@@ -105,7 +122,9 @@
           {
             "name": "teacher",
             "parent": "Catalog",
-            "arguments": { "id": "Int!" },
+            "arguments": {
+              "id": "Int!"
+            },
             "returns": "Teacher",
             "rest_equivalent": "GET /api/catalog/teachers/[id]",
             "mcp_equivalent": "catalog_teacher_get"
