+++ docs/contracts/todo.json
@@ -18,10 +18,14 @@
       "title": "Todo List & Filter",
       "auth": "user",
       "links": {
-        "models": ["Todo"]
+        "models": [
+          "Todo"
+        ]
       },
       "web": {
-        "pages": ["/workspace/todos"]
+        "pages": [
+          "/workspace/todos"
+        ]
       },
       "graphql": {
         "fields": [
@@ -34,7 +38,9 @@
             },
             "returns": "TodoPage!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:read"],
+            "required_scopes": [
+              "workspace.todo:read"
+            ],
             "rest_equivalent": "GET /api/workspace/todos",
             "mcp_equivalent": "workspace_todo_list",
             "notes": [
@@ -45,10 +51,14 @@
         "mutations": [
           {
             "name": "todoCreate",
-            "arguments": { "input": "CreateTodoInput!" },
+            "arguments": {
+              "input": "CreateTodoInput!"
+            },
             "returns": "TodoMutationPayload!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:write"],
+            "required_scopes": [
+              "workspace.todo:write"
+            ],
             "rest_equivalent": "POST /api/workspace/todos",
             "mcp_equivalent": "workspace_todo_create"
           },
@@ -60,16 +70,22 @@
             },
             "returns": "TodoMutationPayload!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:write"],
+            "required_scopes": [
+              "workspace.todo:write"
+            ],
             "rest_equivalent": "PATCH /api/workspace/todos/[id]",
             "mcp_equivalent": "workspace_todo_update"
           },
           {
             "name": "todoDelete",
-            "arguments": { "id": "ID!" },
+            "arguments": {
+              "id": "ID!"
+            },
             "returns": "DeleteMutationPayload!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:write"],
+            "required_scopes": [
+              "workspace.todo:write"
+            ],
             "rest_equivalent": "DELETE /api/workspace/todos/[id]",
             "mcp_equivalent": "workspace_todo_delete"
           },
@@ -80,7 +96,9 @@
             },
             "returns": "TodoCompletionBatchPayload!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:write"],
+            "required_scopes": [
+              "workspace.todo:write"
+            ],
             "rest_equivalent": "PATCH /api/workspace/todos/batch",
             "mcp_equivalent": "workspace_todo_update",
             "notes": [
@@ -90,10 +108,14 @@
           },
           {
             "name": "todosDelete",
-            "arguments": { "ids": "[ID!]!" },
+            "arguments": {
+              "ids": "[ID!]!"
+            },
             "returns": "TodoDeleteBatchPayload!",
             "auth": "user",
-            "required_scopes": ["workspace.todo:write"],
+            "required_scopes": [
+              "workspace.todo:write"
+            ],
             "rest_equivalent": "DELETE /api/workspace/todos/batch",
             "mcp_equivalent": "workspace_todo_delete",
             "notes": [
@@ -146,10 +168,14 @@
       "title": "Todo Create/Edit Form",
       "auth": "user",
       "links": {
-        "models": ["Todo"]
+        "models": [
+          "Todo"
+        ]
       },
       "web": {
-        "pages": ["/workspace/todos"]
+        "pages": [
+          "/workspace/todos"
+        ]
       },
       "rest": {
         "routes": [
