+++ docs/contracts/user.json
@@ -32,7 +32,9 @@
       "title": "Sign In",
       "auth": "anon",
       "web": {
-        "pages": ["/account/sign-in"]
+        "pages": [
+          "/account/sign-in"
+        ]
       },
       "rest": "stable",
       "mcp": "unavailable",
@@ -49,12 +51,16 @@
       "title": "First Login & Welcome Page",
       "auth": "user",
       "web": {
-        "pages": ["/account/welcome"]
+        "pages": [
+          "/account/welcome"
+        ]
       },
       "rest": "unavailable",
       "mcp": "unavailable",
       "links": {
-        "models": ["User"]
+        "models": [
+          "User"
+        ]
       },
       "display": {
         "fields": [
@@ -85,7 +91,9 @@
       "title": "Public Profile",
       "auth": "anon",
       "web": {
-        "pages": ["/community/users/$identifier"]
+        "pages": [
+          "/community/users/$identifier"
+        ]
       },
       "rest": {
         "routes": [
@@ -111,7 +119,9 @@
           {
             "parent": "Community",
             "name": "user",
-            "arguments": { "identifier": "String!" },
+            "arguments": {
+              "identifier": "String!"
+            },
             "returns": "CommunityUser",
             "auth": "anon",
             "rest_equivalent": "GET /api/community/users/:identifier",
@@ -124,7 +134,9 @@
             "name": "profile",
             "returns": "UserProfile!",
             "auth": "user",
-            "required_scopes": ["account.profile:read"],
+            "required_scopes": [
+              "account.profile:read"
+            ],
             "rest_equivalent": "GET /api/account/profile",
             "mcp_equivalent": "account_profile_get"
           }
@@ -149,7 +161,9 @@
         ]
       },
       "links": {
-        "models": ["User"]
+        "models": [
+          "User"
+        ]
       },
       "display": {
         "fields": [
@@ -185,7 +199,9 @@
       "rest": "unavailable",
       "mcp": "unavailable",
       "links": {
-        "models": ["User"]
+        "models": [
+          "User"
+        ]
       },
       "display": {
         "fields": [
@@ -231,7 +247,9 @@
               "cursor": "String",
               "limit": "Int"
             },
-            "required_scopes": ["account.client-activity:read"],
+            "required_scopes": [
+              "account.client-activity:read"
+            ],
             "returns": "AccountClientActivityPage!"
           }
         ]
@@ -247,7 +265,11 @@
           }
         ]
       },
-      "links": { "models": ["AuditLog"] }
+      "links": {
+        "models": [
+          "AuditLog"
+        ]
+      }
     }
   }
 }
