+++ docs/contracts/young-event.json
@@ -20,7 +20,10 @@
       "title": "Second-Classroom Events",
       "auth": "anon",
       "web": {
-        "pages": ["/catalog/young-events", "/catalog/young-events/[youngId]"]
+        "pages": [
+          "/catalog/young-events",
+          "/catalog/young-events/[youngId]"
+        ]
       },
       "rest": {
         "routes": [
@@ -36,7 +39,9 @@
           {
             "path": "/api/catalog/young-events/[youngId]",
             "returns": "single event including rawJson",
-            "notes": ["404 when the youngId is unknown."]
+            "notes": [
+              "404 when the youngId is unknown."
+            ]
           },
           {
             "path": "/api/catalog/young-events/[youngId]/image",
@@ -64,11 +69,15 @@
           {
             "name": "youngEvent",
             "parent": "Catalog",
-            "arguments": { "youngId": "String!" },
+            "arguments": {
+              "youngId": "String!"
+            },
             "returns": "YoungEvent",
             "rest_equivalent": "GET /api/catalog/young-events/[youngId]",
             "mcp_equivalent": "catalog_young_event_get",
-            "notes": ["Returns null when the youngId is unknown."]
+            "notes": [
+              "Returns null when the youngId is unknown."
+            ]
           }
         ]
       },
@@ -87,7 +96,9 @@
             "name": "catalog_young_event_get",
             "returns": "single event including rawJson",
             "rest_equivalent": "GET /api/catalog/young-events/[youngId]",
-            "notes": ["Accepts youngId and mode inputs."]
+            "notes": [
+              "Accepts youngId and mode inputs."
+            ]
           }
         ]
       },
