+++ src/lib/mcp/tools/workspace/dashboard-tool-snapshot.ts
@@ -1,4 +1,4 @@
-import { getAssistantDashboardSnapshot } from "@/features/workspace/server/assistant-dashboard-snapshot";
+import { getAssistantDashboardSnapshot } from "@/features/workspace/server/workspace-snapshot";
 import type { AppLocale } from "@/i18n/config";
 import {
   getUserId,
