+++ tests/e2e/src/app/workspace/exams/test.ts
@@ -75,7 +75,7 @@ test.describe("仪表盘考试", () => {
 
   test("移动端考试工具栏直接筛选并保持卡片视图", async ({ page }, testInfo) => {
     await page.addInitScript(() => {
-      localStorage.removeItem("life-ustc-dashboard-view-mode");
+      localStorage.removeItem("life-ustc-workspace-view-mode");
     });
     await page.setViewportSize({ height: 844, width: 390 });
     await signInAsDebugUser(page, "/workspace/exams");
